+++ src/main/java/com/erp/Inventory/Management/model/Product.java
@@ -11,8 +11,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
     private String sku;
     private Double price;
@@ -22,11 +21,11 @@ public class Product {
     @JoinColumn(name = "category_id")
     private Category category;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
